requirements.txt
tiktoken
unstructured
chromadb
openai
pdfplumber
python-dotenv
requests
transformers
langchain
langchain-community
multilingual-pdf2text
nltk
torch
urllib3
typer==0.9.4
spacy<3.8.0
fastapi<0.111.0
accelerate
pyngrok
